# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "battery life — API для управления заявками расчёта времени работы аккумуляторов",
        "title": "battery life API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/battery_life_item/add/{battery_life_type_id}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Добавляет аккумулятор в черновик пользователя. При отсутствии черновика он создается.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "battery life"
                ],
                "summary": "Добавить аккумулятор в заявку",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID типа аккумулятора",
                        "name": "battery_life_type_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Параметры позиции",
                        "name": "item",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/RIP2026_internal_app_serializer.BatteryLifeItemJSON"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/RIP2026_internal_app_serializer.BatteryLifeJSON"
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/RIP2026_internal_app_serializer.BatteryLifeJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/battery_life_item/{battery_life_type_id}/{battery_life_id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Обновляет ток и количество для позиции в черновике заявки.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "battery life"
                ],
                "summary": "Изменить позицию в заявке",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID типа аккумулятора",
                        "name": "battery_life_type_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "battery_life_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные позиции",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RIP2026_internal_app_serializer.BatteryLifeItemJSON"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/RIP2026_internal_app_serializer.BatteryLifeItemJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удаляет позицию из черновика заявки.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "battery life"
                ],
                "summary": "Удалить аккумулятор из заявки",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID типа аккумулятора",
                        "name": "battery_life_type_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "battery_life_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/RIP2026_internal_app_serializer.BatteryLifeJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "RIP2026_internal_app_serializer.BatteryLifeItemJSON": {
            "type": "object",
            "properties": {
                "battery_life_id": {
                    "type": "integer"
                },
                "battery_type_id": {
                    "type": "integer"
                },
                "current_ma": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "runtime_hours": {
                    "type": "number"
                }
            }
        },
        "RIP2026_internal_app_serializer.BatteryLifeJSON": {
            "type": "object",
            "properties": {
                "completed_at": {
                    "type": "string"
                },
                "completed_item_count": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "creator_login": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "formed_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "moderator_login": {
                    "description": "опционально, как в SystemLoadToJSON",
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "total_runtime_hours": {
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}